add zuluru sync to cli
